requirements.txt
# Core Python dependencies
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
Jinja2==3.1.4
# Database
aiosqlite==0.20.0
# Testing
pytest==8.3.3
respx==0.21.1
